% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Define a function to create a scatter plot}
\usage{
create_scatter_plot(data, x_column, y_column, title)
}
\arguments{
\item{title}{the title of the scatter plot}

\item{dataframe}{the data frame to plot}

\item{x_col}{the x-column of the data frame}

\item{y_col}{the y-column of the data frame}
}
\value{
the scatter plot diagram
}
\description{
Define a function to create a scatter plot
}
\examples{
create_scatter_plot(df, x='weight', y='age', title='scatter plot')
}
