% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{estegn}
\alias{estegn}
\title{function performing statistical analysis to estimate the parameters
of a normal distribution for a given data set along with a confidence intervals
for the mean and standard deviation.}
\usage{
estegn(x, alpha = 0.05)
}
\arguments{
\item{x}{Input data set}

\item{alpha}{significance level}
}
\value{
Return a list containing the following elements: G, V, Cov, StdErr, ConfInts
}
\description{
function performing statistical analysis to estimate the parameters
of a normal distribution for a given data set along with a confidence intervals
for the mean and standard deviation.
}
\examples{
x = c(2, 4, 6, 8, 10, 30, 45, 55)
estegn(x, alpha = 0.01)
}
