% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{read_data_from_csv}
\alias{read_data_from_csv}
\title{Define a function to read data from a CSV file}
\usage{
read_data_from_csv(file_path)
}
\arguments{
\item{file_path}{read the file path to the csv file}
}
\value{
the dataframe of the input file
}
\description{
Define a function to read data from a CSV file
}
\examples{
df = read_data_from_csv("data.csv")
}
