% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{solve_quadratic}
\alias{solve_quadratic}
\title{function to solve the quadratic equation}
\usage{
solve_quadratic(a, b, c)
}
\arguments{
\item{a}{value for a}

\item{b}{value for b}

\item{c}{value for c}
}
\description{
function to solve the quadratic equation
}
\examples{
a <- 1; b <- -3; c <- 4
roots <- solve_quadratic(a, b, c)
cat("Roots:", roots, "\n")

a <- 2; b <- -3; c <- 1
roots <- solve_quadratic(a, b, c)
cat("Roots:", roots, "\n")

}
