% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{create_scatter_plot}
\alias{create_scatter_plot}
\title{Define a function to create a scatter plot}
\usage{
create_scatter_plot(data, x_column, y_column, title)
}
\arguments{
\item{title}{the title of the scatter plot}

\item{dataframe}{the dataframe to plot}

\item{x_col}{the x-column of the dataframe}

\item{y_col}{the y-column of the dataframe}
}
\value{
the scatterplot diagram
}
\description{
Define a function to create a scatter plot
}
\examples{
create_scatter_plot(df, x=df['weight'], y=df['age'], title='scatter plot')
}
